修改很多的细节

+++ ZFB/Class/Main/Home/Model/AllStoreModel.h
@@ -68,7 +68,7 @@
 
 @property (nonatomic, assign) NSInteger storeId;
 
-@property (nonatomic, assign) CGFloat starLevel;
+@property (nonatomic, assign) NSInteger starLevel;
 
 @property (nonatomic, assign) NSInteger storekeeperUserId;
